refactor(warehouse): move warehouse-owned code out of core
Four modules under posthog/ were pure warehouse_sources logic that happened to
live in core, so every one of them had to reach back through the facade for the
models it operates on. Moving them into the product turns those cross-boundary
imports into ordinary internal ones.

- The column-validation celery task reads a table, calls the model's own
  validate_column_type and saves it. Its name is pinned to the old dotted path
  so a message queued before the deploy still routes.
- The activity-log detail helpers had exactly one caller, the product's own
  activity_logging module.
- Two management commands operate only on schemas. Django finds commands by app,
  so the invocation is unchanged.

Tests move with their subjects, including the patch targets that named the old
module paths.
