+++ products/data_warehouse/backend/presentation/views/table.py
@@ -22,7 +22,6 @@ from posthog.models import Team
 from posthog.models.user import User
 from posthog.rbac.access_control_api_mixin import AccessControlViewSetMixin
 from posthog.rbac.user_access_control import UserAccessControlSerializerMixin
-from posthog.tasks.warehouse import validate_data_warehouse_table_columns
 
 from products.data_warehouse.backend.facade.api import get_s3_client
 from products.warehouse_sources.backend.facade.api import (
@@ -45,6 +44,7 @@ from products.warehouse_sources.backend.facade.models import (
     ExternalDataSource,
     validate_warehouse_table_url_pattern,
 )
+from products.warehouse_sources.backend.facade.tasks import validate_data_warehouse_table_columns
 from products.warehouse_sources.backend.presentation.views.external_data_source import (
     SimpleExternalDataSourceSerializers,
 )
@@ -809,7 +809,7 @@ class TableViewSet(TeamAndOrgViewSetMixin, AccessControlViewSetMixin, viewsets.M
                 table.save(internally_computed_url_pattern=True)
 
                 # Validate columns in background
-                from posthog.tasks.warehouse import validate_data_warehouse_table_columns
+                from products.warehouse_sources.backend.facade.tasks import validate_data_warehouse_table_columns
 
                 validate_data_warehouse_table_columns.delay(team_id, str(table.id))
