+++ products/data_warehouse/backend/tests/api/test_table.py
@@ -145,7 +145,7 @@ class TestTable(APIBaseTest):
         "products.warehouse_sources.backend.models.table.DataWarehouseTable.validate_column_type",
         return_value=True,
     )
-    @patch("posthog.tasks.warehouse.get_client")
+    @patch("products.warehouse_sources.backend.tasks.tasks.get_client")
     def test_create_columns(self, patch_get_columns, patch_validate_column_type, patch_get_client):
         response = self.client.post(
             f"/api/projects/{self.team.id}/warehouse_tables/",
@@ -186,7 +186,7 @@ class TestTable(APIBaseTest):
         "products.warehouse_sources.backend.models.table.DataWarehouseTable.validate_column_type",
         return_value=False,
     )
-    @patch("posthog.tasks.warehouse.get_client")
+    @patch("products.warehouse_sources.backend.tasks.tasks.get_client")
     def test_create_columns_invalid_schema(self, patch_get_columns, patch_validate_column_type, patch_get_client):
         response = self.client.post(
             f"/api/projects/{self.team.id}/warehouse_tables/",
@@ -379,7 +379,7 @@ class TestTable(APIBaseTest):
         "products.warehouse_sources.backend.models.table.DataWarehouseTable.validate_column_type",
         return_value=True,
     )
-    @patch("posthog.tasks.warehouse.get_client")
+    @patch("products.warehouse_sources.backend.tasks.tasks.get_client")
     def test_table_name_duplicate(self, patch_get_columns, patch_validate_column_type, patch_get_client):
         response = self.client.post(
             f"/api/projects/{self.team.id}/warehouse_tables/",
