+++ products/data_warehouse/backend/tests/api/test_view.py
@@ -102,7 +102,7 @@ class TestView(APIBaseTest):
         "products.data_modeling.backend.models.datawarehouse_saved_query.DataWarehouseSavedQuery.get_columns",
         return_value={"id": "String", "a_column": "String"},
     )
-    @patch("posthog.tasks.warehouse.get_client")
+    @patch("products.warehouse_sources.backend.tasks.tasks.get_client")
     def test_view_with_external_table(self, patch_get_columns_1, patch_get_columns_2, patch_get_client):
         response = self.client.post(
             f"/api/environments/{self.team.id}/warehouse_tables/",
