+++ products/warehouse_sources/backend/activity_logging_utils.py
@@ -1,6 +1,6 @@
 from posthog.models.activity_logging.utils import CreatedByInfo
 
-from products.warehouse_sources.backend.facade.models import ExternalDataSource
+from products.warehouse_sources.backend.models.external_data_source import ExternalDataSource
 
 
 def get_external_data_source_detail_name(external_data_source: ExternalDataSource) -> str:
