+++ products/warehouse_sources/backend/facade/tasks.py
@@ -1,5 +1,5 @@
 """Celery tasks core schedules for warehouse_sources (see products/architecture.md, wiring couplings)."""
 
-from products.warehouse_sources.backend.tasks import sweep_stopped_schema_syncs
+from products.warehouse_sources.backend.tasks import sweep_stopped_schema_syncs, validate_data_warehouse_table_columns
 
-__all__ = ["sweep_stopped_schema_syncs"]
+__all__ = ["sweep_stopped_schema_syncs", "validate_data_warehouse_table_columns"]
