+++ products/warehouse_sources/backend/management/commands/sync_import_pipeline_schedules.py
@@ -1,7 +1,7 @@
 from django.core.management.base import BaseCommand
 
 from products.data_warehouse.backend.facade.api import external_data_workflow_exists, sync_external_data_job_workflow
-from products.warehouse_sources.backend.facade.models import ExternalDataSchema
+from products.warehouse_sources.backend.models.external_data_schema import ExternalDataSchema
 
 
 class Command(BaseCommand):
