+++ products/warehouse_sources/backend/tasks/__init__.py
@@ -1,4 +1,5 @@
 from products.warehouse_sources.backend.tasks.tasks import (  # noqa: F401
     cleanup_disabled_external_data_schema,
     sweep_stopped_schema_syncs,
+    validate_data_warehouse_table_columns,
 )
