+++ products/warehouse_sources/backend/tests/test_validate_table_columns.py
@@ -1,13 +1,12 @@
 from posthog.test.base import APIBaseTest
 from unittest.mock import MagicMock, patch
 
-from posthog.tasks.warehouse import validate_data_warehouse_table_columns
+from products.warehouse_sources.backend.models.table import DataWarehouseTable
+from products.warehouse_sources.backend.tasks.tasks import validate_data_warehouse_table_columns
 
-from products.warehouse_sources.backend.facade.models import DataWarehouseTable
 
-
-class TestWarehouse(APIBaseTest):
-    @patch("posthog.tasks.warehouse.get_client")
+class TestValidateDataWarehouseTableColumns(APIBaseTest):
+    @patch("products.warehouse_sources.backend.tasks.tasks.get_client")
     def test_validate_data_warehouse_table_columns(self, mock_get_client: MagicMock) -> None:
         mock_ph_client = MagicMock()
         mock_get_client.return_value = mock_ph_client
